moved cardtest() to cardtest.rs
